Create db.models.roles.py
